+++ yx_scope/packages/yx_scope/lib/src/core/scope_state_holder.dart
@@ -139,6 +139,10 @@ class ScopeStateHolder<Scope> {
     }(), '');
     return true;
   }
+
+  @override
+  String toString() =>
+      '${objectRuntimeType(this, 'ScopeStateHolder')}(state: $_state)';
 }
 
 @visibleForTesting